# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-providers:
    get:
      operationId: getServiceProviders
      tags: [providers]
      summary: List genuinely approved providers
      parameters:
        - name: serviceId
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        "200":
          description: Approved providers, optionally filtered by supported service
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServiceProvider"
  /service-providers/{providerId}:
    get:
      operationId: getServiceProvider
      tags: [providers]
      summary: Get one genuinely approved provider
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceProvider"
        "404":
          description: Provider not found or not approved
  /provider/profile:
    get:
      operationId: getProviderProfile
      tags: [provider-account]
      summary: Get the signed-in user's provider application or settings
      responses:
        "200":
          description: Provider profile, or null before one is created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderProfileEnvelope"
        "401":
          description: Not signed in
    put:
      operationId: updateProviderProfile
      tags: [provider-account]
      summary: Create or update provider application/settings fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderProfileEdit"
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderProfileEnvelope"
        "400":
          description: Invalid profile
        "401":
          description: Not signed in
  /provider/profile/submit:
    post:
      operationId: submitProviderProfile
      tags: [provider-account]
      summary: Submit or resubmit a provider application for admin review
      responses:
        "200":
          description: Submitted profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderProfileEnvelope"
        "401":
          description: Not signed in
        "404":
          description: Profile not found
        "409":
          description: Profile cannot be submitted
  /provider/orders:
    get:
      operationId: getProviderOrders
      tags: [provider-account]
      summary: List service requests assigned to the approved owner
      responses:
        "200":
          description: Assigned requests; exact location/photos are exposed only while accepted or after completion
          content:
            application/json:
              schema:
                type: object
                required: [orders]
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProviderOrder"
        "401":
          description: Not signed in
        "403":
          description: Approved owner role required
  /provider/orders/{id}:
    patch:
      operationId: respondToProviderOrder
      tags: [provider-account]
      summary: Atomically respond to pending work or terminally update accepted work
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderResponseInput"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: "#/components/schemas/ProviderOrder"
        "400":
          description: Invalid decision or missing required reason
        "401":
          description: Not signed in
        "403":
          description: Owner role required
        "409":
          description: Request is not assigned work in the required prior state, or another transition won
...
components:
  schemas:
...
    OrderKind:
      type: string
      enum: [service, product]
    OrderStatus:
      type: string
      enum: [pending, accepted, rejected, completed, cancelled]
    CancellationActor:
      type: string
      enum: [customer, provider]
    OrderItem:
      type: object
      additionalProperties: false
      required: [id, name, quantity]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 200
        name:
          type: string
          minLength: 1
          maxLength: 200
        quantity:
          type: integer
          minimum: 1
          maximum: 1000
    ServiceProvider:
      type: object
      additionalProperties: false
      required: [id, name, serviceIds, servicePrices, rating, distanceKm, earliestAvailableAt, isDemo]
      properties:
        id:
          type: string
        name:
          type: string
        serviceIds:
          type: array
          items:
            type: string
        servicePrices:
          $ref: "#/components/schemas/ServicePrices"
        rating:
          type: ["number", "null"]
          minimum: 0
          maximum: 5
        distanceKm:
          type: ["number", "null"]
          minimum: 0
        earliestAvailableAt:
          type: ["string", "null"]
        imageUrl:
          type: string
        isDemo:
          type: boolean
    ServiceBookingNeeds:
      type: object
      additionalProperties: false
      properties:
        approximateAreaM2:
          type: number
          exclusiveMinimum: 0
        shrubCount:
          type: integer
          minimum: 1
        standingTreeCount:
          type: integer
          minimum: 1
        stumpOrRootAreaCount:
          type: integer
          minimum: 1
        treeAccessDetails:
          type: string
          minLength: 1
          maxLength: 2000
...
    ServicePrices:
      type: object
      additionalProperties:
        type: ["number", "null"]
        minimum: 0
    ProviderStatus:
      type: string
      enum: [draft, pending, approved, rejected, suspended]
    ProviderProfileEdit:
      type: object
      additionalProperties: false
      required: [shopName, serviceIds, description, earliestAvailableAt, servicePrices]
      properties:
        shopName:
          type: string
          minLength: 2
          maxLength: 150
        serviceIds:
          type: array
          minItems: 1
          maxItems: 50
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 5000
        earliestAvailableAt:
          type: ["string", "null"]
          format: date-time
        servicePrices:
          $ref: "#/components/schemas/ServicePrices"
    ProviderProfile:
      allOf:
        - $ref: "#/components/schemas/ProviderProfileEdit"
        - type: object
          required: [id, status, statusReason, submittedAt, reviewedAt, isDemo, createdAt, updatedAt]
          properties:
            id:
              type: string
            status:
              $ref: "#/components/schemas/ProviderStatus"
            statusReason:
              type: ["string", "null"]
            submittedAt:
              type: ["string", "null"]
              format: date-time
            reviewedAt:
              type: ["string", "null"]
              format: date-time
            isDemo:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    ProviderProfileEnvelope:
      type: object
      required: [profile]
      properties:
        profile:
          oneOf:
            - $ref: "#/components/schemas/ProviderProfile"
            - type: "null"
...
    ProviderBooking:
      type: object
      additionalProperties: false
      required: [location, taskIds, preferredTime, photoPaths]
      properties:
        gardenName:
          type: string
        location:
          type: ["string", "null"]
        taskIds:
          type: array
          items:
            type: string
        notes:
          type: string
        needs:
          $ref: "#/components/schemas/ServiceBookingNeeds"
        preferredTime:
          type: string
        photoPaths:
          type: array
          items:
            type: string
    ProviderOrder:
      type: object
      required: [id, kind, items, status, responseReason, respondedAt, cancellationReason, cancelledAt, cancelledBy, completedAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/OrderKind"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        details:
          type: ["string", "null"]
        booking:
          oneOf:
            - $ref: "#/components/schemas/ProviderBooking"
            - type: "null"
        status:
          $ref: "#/components/schemas/OrderStatus"
        responseReason:
          type: ["string", "null"]
        respondedAt:
          type: ["string", "null"]
          format: date-time
        cancellationReason:
          type: ["string", "null"]
        cancelledAt:
          type: ["string", "null"]
          format: date-time
        cancelledBy:
          oneOf:
            - $ref: "#/components/schemas/CancellationActor"
            - type: "null"
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    OrderResponseInput:
      oneOf:
        - type: object
          additionalProperties: false
          required: [status]
          properties:
            status:
              type: string
              enum: [accepted, completed]
            reason:
              type: string
              maxLength: 2000
        - type: object
          additionalProperties: false
          required: [status, reason]
          properties:
            status:
              type: string
              const: rejected
            reason:
              type: string
              minLength: 1
              maxLength: 2000
        - type: object
          additionalProperties: false
          required: [status, reason]
          properties:
            status:
              type: string
              const: cancelled
            reason:
              type: string
              minLength: 1
              maxLength: 1000
...