# lib/api-zod/src/generated/types/serviceProvider.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServicePrices } from './servicePrices';

export interface ServiceProvider {
  id: string;
  name: string;
  serviceIds: string[];
  servicePrices: ServicePrices;
  /**
     * @minimum 0
     * @maximum 5
     * @nullable
     */
  rating: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  distanceKm: number | null;
  /** @nullable */
  earliestAvailableAt: string | null;
  imageUrl?: string;
  isDemo: boolean;
}